Use a more complex function to shorten a URL
Previously, to shorten a URL we were using the hashcode of the URL as the corresponding
"shortened" URL. However, hashcodes are not unique, and we may run into collisions.

Now, we take the original URL and compute its SHA (which is bound to be unique for a given
string). There is a global counter which is incremented every time a new SHA is encountered.
We create a mapping between the base64-encoded counter and the sha, as well as the
base64-encoded counter and the original URL. The base64-encoded counter is now treated as the
"shortened" url.

Every time a request is made to shorten a URL, we first perform a look up to see if such a
SHA already exists in the storage. If it does, we return the previously computed shortened
url. Otherwise, we compute the shortened url and return that.